Inserted a sound table + new function
Made the table for easier management of the sounds. Also created a new function that will get the size of the sound table - since it's going to increase over time, we won't have to keep changing things in different places. This will also be used to select a random index of the table to play a random sound.

Currently the new function is causing the main bit of code to not work. Need to investigate.