Important changes
Fixed issue where I was including two wow addon namespaces, where only one is allowed. This caused the code to not run at all and was confusing me for a while.
Made the namespace Soundboard. It's just a global table that is local to your addon. Also, renamed the SoundTable.lua file to just Sounds.lua. SoundTable.lua is deprecrated.
Utils.lua I have been using for a while - but with the recent namespace change - I can upload it. Just some commands.